Remove root.go
